+++ lib/XmlReporter.js
@@ -1,0 +1,83 @@
+/*
+ * grunt-qunit-junit
+ * https://github.com/sbrandwoo/grunt-qunit-junit
+ *
+ * Copyright (c) 2012 Stephen Brandwood
+ * Licensed under the MIT license.
+ */
+
+'use strict';
+
+var _ = require('underscore');
+
+function XmlReporter() {
+
+}
+_.extend(XmlReporter.prototype, {
+    escape: function (value) {
+        return value.replace(/&/g, '&amp;')
+            .replace(/</g, '&lt;')
+            .replace(/>/g, '&gt;')
+            .replace(/"/g, '&quot;');
+    },
+
+    generateReport: function (result) {
+        var xml = '<?xml version="1.0" encoding="UTF-8"?>\n<testsuites>\n';
+        _.each(result.modules, function (module) {
+            xml += '\t<testsuite'
+                + ' name="' + this.escape(result.classname) + '"'
+                + ' errors="' + module.errored + '"'
+                + ' failures="' + module.failed + '"'
+                + ' tests="' + module.tests.length + '"'
+                + ' time="0.01">\n';
+            _.each(module.tests, function (test) {
+                xml += '\t\t<testcase'
+                    + ' classname="' + this.escape(result.classname) + '"'
+                    + ' name="' + this.escape(
+                        (module.name ? (module.name + ": ") : "") + test.name) + '"'
+                    + ' assertions="' + test.total + '" time="0.01">\n';
+                _.each(test.logs, function (data) {
+                    xml += '\t\t\t<' + data.type + ' type="failed" message="'
+                        + this.escape(data.message) + '">\n';
+                    if (data.stack) {
+                        xml += '\t' + this.escape(data.stack) + '\n';
+                    }
+                    if (data.source){
+                        xml += '\t' + this.escape(data.source) + '\n';
+                    }
+                    xml += '\t\t\t</' + data.type + '>\n';
+                }, this);
+                xml += "\t\t</testcase>\n";
+            }, this);
+            xml += "\t</testsuite>\n";
+        }, this);
+        xml += "</testsuites>\n";
+
+        return xml;
+    },
+
+    generateTimeout: function (result) {
+        var xml = '<?xml version="1.0" encoding="UTF-8"?>\n<testsuites>\n';
+
+        xml += '\t<testsuite'
+            + ' name="' + this.escape(result.classname) + '"'
+            + ' errors="1"'
+            + ' failures="0"'
+            + ' tests="1"'
+            + ' time="0.01">\n';
+        xml += '\t\t<testcase'
+            + ' classname="' + this.escape(result.classname) + '"'
+            + ' name="main"'
+            + ' assertions="1"'
+            + ' time="0.01">\n';
+        xml += '\t\t\t<error type="timeout" message="Test timed out, '
+            + 'possibly due to a missing QUnit.start() call."></error>\n';
+        xml += "\t\t</testcase>\n";
+        xml += "\t</testsuite>\n";
+        xml += "</testsuites>\n";
+
+        return xml;
+    }
+});
+
+module.exports = XmlReporter;
